requirements.txt
PyYAML==6.0.2
requests==2.32.5
websockets==15.0.1